requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==21.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2
django-allauth==0.44.0
django-comments-xtd==2.9.0
django-contrib-comments==2.1.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-jalali-date==0.3.1
django-ranged-response==0.2.0
django-resized==0.3.11
django-simple-captcha==0.5.14
djangorestframework==3.12.4
docutils==0.17
idna==2.10
isodate==0.6.0
jdatetime==3.6.2
lxml==4.6.3
oauthlib==3.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
testfixtures==6.17.1
urllib3==1.26.4
zeep==4.0.0
